code/PythonApp/requirements.txt
opencv-python
pillow
scikit-image
torch
